requirements.txt
google-adk
yfinance
pandas
numpy
matplotlib
mplfinance
pydantic
python-dotenv
pytest
